+++ src/main/kotlin/com/github/jensim/megamanipulator/http/jetbrainsmarketplace/PluginVersionResponse.kt
@@ -1,0 +1,5 @@
+package com.github.jensim.megamanipulator.http.jetbrainsmarketplace
+
+data class PluginVersionResponse(
+    val version: String,
+)