+++ src/main/kotlin/com/github/jensim/megamanipulator/toolswindow/MegaManipulatorTabContentCreator.kt
@@ -1,6 +1,7 @@
 package com.github.jensim.megamanipulator.toolswindow
 
 import com.github.jensim.megamanipulator.MyBundle
+import com.github.jensim.megamanipulator.actions.ProcessOperator
 import com.github.jensim.megamanipulator.actions.apply.ApplyWindow
 import com.github.jensim.megamanipulator.actions.forks.ForksWindow
 import com.github.jensim.megamanipulator.actions.git.CloneHistoryWindow
@@ -8,6 +9,7 @@ import com.github.jensim.megamanipulator.actions.git.GitWindow
 import com.github.jensim.megamanipulator.actions.search.SearchWindow
 import com.github.jensim.megamanipulator.actions.vcs.PullRequestWindow
 import com.github.jensim.megamanipulator.files.FilesOperator
+import com.github.jensim.megamanipulator.http.jetbrainsmarketplace.JetbrainsMarketplaceClient
 import com.github.jensim.megamanipulator.settings.SettingsWindow
 import com.github.jensim.megamanipulator.ui.DialogGenerator
 import com.intellij.openapi.components.service
@@ -23,6 +25,7 @@ import com.intellij.ui.content.ContentFactory
 import com.intellij.ui.content.ContentManagerEvent
 import com.intellij.ui.content.ContentManagerListener
 import com.intellij.util.ui.components.BorderLayoutPanel
+import kotlinx.coroutines.runBlocking
 import org.slf4j.LoggerFactory
 import java.io.File
 import javax.swing.JButton
@@ -135,11 +138,74 @@ class MegaManipulatorTabContentCreator(
         } catch (e: Exception) {
             logger.warn("Was unable to refresh the settings page")
         }
+
+        validatePath(project)
+        validatePluginVersion(project, filesOperator)
     }
 
     override fun tabSelectionRequested(tabKey: TabKey) {
         selectMap[tabKey]?.let {
             toolWindow?.contentManager?.setSelectedContent(it)
+        }
+    }
+
+    private fun validatePath(project: Project) {
+        val osProperty = System.getProperty("os.name").lowercase()
+        if (osProperty.startsWith("darwin") || osProperty.startsWith("mac os x")) {
+            val output: List<String> = checkCommands(listOf("brew", "git", "find"), project.service())
+            if (output.isNotEmpty()) {
+                val dialoger = project.service<DialogGenerator>()
+                dialoger.showConfirm(
+                    title = "Borked PATH, probably.",
+                    yesText = "Ok",
+                    noText = "Cancel",
+                    focusComponent = null,
+                    onYes = {},
+                    message = """
+                        Cannot detect the following commands on the system PATH: $output<br>
+                        You must open IntelliJ from a command line, or the Java process will have a totally borked PATH.<br>
+                        This will affect you when you do scripted changes using the Apply tab.<br>
+                        Try closing all IntelliJ windows, and restart it one of these ways (from the <u>terminal</u>):<br>
+                        <ul>
+                        <li style="border: dashed;">
+                        <pre>$ open -a IntelliJ\ IDEA\ Ultimate</pre>
+                        </li>
+                        <li style="border: dashed;">
+                        <pre>$ idea</pre>
+                        If you have the idea toolbox and shellscript installed
+                        </li>
+                        </ul>
+                    """.trimIndent()
+                )
+            } else {
+                logger.info("All seems okay with the path")
+            }
+        }
+    }
+
+    private fun checkCommands(commands: List<String>, processOperator: ProcessOperator): List<String> {
+        return runBlocking {
+            commands.mapNotNull {
+                val result = processOperator.runCommandAsync(File("/"), listOf("command", "-v", it)).await()
+                if (result.exitCode != 0) it else null
+            }
+        }
+    }
+
+    private fun validatePluginVersion(project: Project, filesOperator: FilesOperator) {
+        filesOperator.getPluginVersion()?.let { pluginVersion ->
+            JetbrainsMarketplaceClient().getLatestVersion()?.let { latestVersion ->
+                if (pluginVersion != latestVersion) {
+                    val dialoger = project.service<DialogGenerator>()
+                    dialoger.showConfirm(
+                        title = "Plugin version outdated!",
+                        message = "A new plugin version is available, please update ;-)",
+                        focusComponent = null,
+                        yesText = "Yes, boss!",
+                        noText = "MAKE ME!!!",
+                    ) {}
+                }
+            }
         }
     }
 }